requirements.txt
google-genai
pydantic
pandas
beautifulsoup4
python-dotenv
